+++ lib/node-radius/types/ICommon.ts
