+++ lib/node-radius/types/IRadius.ts
