+++ lib/node-radius/types/IResponse.ts
